+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Do case-insensitive file compare when smart cleaning (#203)
+- Fix unwrapping Zod schemas for nullable record structs (#207)
 
 ### Changed